LIBRARY: minor tweaks
